Added background image to a canvas
This is to make the background re-scalable to screen size. Though I don't like it stretching so much.

It is on a separate canvas to the rest of the UI so that the background can be rendered behind the board objects in world space. This is done by the Screen Space - Camera option in the Canvas component