Fixed the dictionary argument exception
Fixed it by shifting the moving code for the furthest piece (i.e. the piece with the longest travel) to get performed after all other pieces.
This is so that the furthest piece doesn't land on another piece before that piece gets moved. Those duplicate coordinates is what caused the exception